+++ rust/cli/CHANGELOG.md
@@ -3,6 +3,8 @@
 ## v0.1.8 - 2026-??-??
 
 - Re-added 23QL384.
+- Add `onerom chips` to show supported chip types and their aliases.
+- Fixed `onerom program --scan-slots`, to correctly re-query One ROM after programming.
 
 ## v0.1.7 - 2026-05-18